Final Commit
I finalized the tkinter GUI. I added buttons that update the vehicle and status message. This took a while to figure out and research but it is working! I also adjusted the size so that it all fits when opened. Give it a shot and let me know what you think!